requirements.txt
fastapi==0.115.0
langchain_community==0.3.0
langchain_core==0.3.15
langchain_google_genai==2.0.6
numpy==1.26.0
google-generativeai==0.8.3
Pillow==10.4.0
pydantic==2.9.2
python_docx==1.1.2
Requests==2.32.3
streamlit==1.36.0
streamlit_drawable_canvas==0.9.3
uvicorn==0.30.6
pyinstaller
python-dotenv